sourcery suggestions needs testing

+++ EmbyArrSync.py
@@ -61,11 +61,9 @@ def get_series_info(user_id):
     }
     response = requests.get(url, params=params)
     if response.status_code == 200:
-        watched_items = response.json()
-        return watched_items
-    else:
-        print(f"Error fetching series info from Emby: {response.status_code}, {response.text}")
-        return None
+        return response.json()
+    print(f"Error fetching series info from Emby: {response.status_code}, {response.text}")
+    return None
     
 def get_watched_items(user_id):
     url = f"{EMBY_URL}/Users/{user_id}/Items/Latest"
@@ -81,27 +79,22 @@ def get_watched_items(user_id):
     }
     response = requests.get(url, params=params)
     if response.status_code == 200:
-        watched_items = response.json()
-        return watched_items
-    else:
-        print(f"Error fetching watched items from Emby: {response.status_code}, {response.text}")
-        return None
+        return response.json()
+    print(f"Error fetching watched items from Emby: {response.status_code}, {response.text}")
+    return None
     
 def get_tmdb_id(movie_name, movie_path):
-    # Try to extract TMDBID from the path
-    match = re.search(r'tmdbid-(\d+)', movie_path)
-    if match:
-        return match.group(1) # Returns the first matching group (the ID)
+    if match := re.search(r'tmdbid-(\d+)', movie_path):
+        return match[1]
     # If the ID is not found in the path, proceed to search via the API
-    search_url = f'https://api.themoviedb.org/3/search/movie'
+    search_url = 'https://api.themoviedb.org/3/search/movie'
     params = {
         'api_key': TMDB_API_KEY,
         'query': movie_name
     }
     response = requests.get(search_url, params=params)
     if response.status_code == 200:
-        results = response.json().get('results', [])
-        if results:
+        if results := response.json().get('results', []):
             # Assuming you want the first result
             first_result = results[0]
             return first_result.get('id')
@@ -119,8 +112,7 @@ def get_tvdb_token():
     response = requests.post(url, json=data, headers=headers)
     if response.status_code == 200:
         token_response = response.json()
-        token = token_response.get('data', {}).get('token')
-        return token
+        return token_response.get('data', {}).get('token')
     else:
         print(f"Failed to obtain TVDB token: {response.status_code}, {response.text}")
         return None
@@ -128,10 +120,8 @@ def get_tvdb_token():
 TVDB_TOKEN = get_tvdb_token()
 
 def get_tvdb_id(series_name, series_path):
-    # Try to extract TVDBID from the path
-    match = re.search(r'tvdbid-(\d+)', series_path)
-    if match:
-        return match.group(1)  # Returns the first matching group (the ID)
+    if match := re.search(r'tvdbid-(\d+)', series_path):
+        return match[1]
     # If the ID is not found in the path, proceed to search via the API
     search_url = f"https://api4.thetvdb.com/v4/search?query={series_name}"
     headers = {
@@ -142,16 +132,14 @@ def get_tvdb_id(series_name, series_path):
     response = requests.get(search_url, headers=headers, params=params)
 
     if response.status_code == 200:
-        results = response.json().get('data', [])
-        if results:
+        if results := response.json().get('data', []):
             # Assuming the first result is what you want
             return results[0].get('tvdb_id')
-        else:
-            print(f"No results found for {series_name}")
-            return None
+        print(f"No results found for {series_name}")
     else:
         print(f"Failed to search TVDB for {series_name}: {response.status_code}, {response.text}")
-        return None
+
+    return None
 
 def get_episode_info(series_id, season_number):
     url = f"{SONARR_URL}/episode"
@@ -164,9 +152,8 @@ def get_episode_info(series_id, season_number):
     response = requests.get(url, params=params)
     if response.status_code == 200:
         return response.json()
-    else:
-        print(f"Error fetching episode info: {response.status_code}, {response.text}")
-        return None
+    print(f"Error fetching episode info: {response.status_code}, {response.text}")
+    return None
 
 def get_series_id_by_tvdb(tvdb_id):
     url = f"{SONARR_URL}/series"
@@ -177,8 +164,7 @@ def get_series_id_by_tvdb(tvdb_id):
     }
     response = requests.get(url, params=params)
     if response.status_code == 200:
-        series_info = response.json()
-        if series_info:
+        if series_info := response.json():
             return series_info[0]['id']  # Assuming the first series is the correct one
     else:
         print(f"Error fetching series by TVDB ID: {response.status_code}, {response.text}")
@@ -197,13 +183,11 @@ def fetch_episode_status(episode_id):
 
 def unmonitor_episodes(episode_ids):
     """Unmonitor episodes if they are currently monitored."""
-    episodes_to_unmonitor = []
-
-    # First, check if each episode is monitored
-    for episode_id in episode_ids:
-        if fetch_episode_status(episode_id):
-            episodes_to_unmonitor.append(episode_id)
-
+    episodes_to_unmonitor = [
+        episode_id
+        for episode_id in episode_ids
+        if fetch_episode_status(episode_id)
+    ]
     if not episodes_to_unmonitor:
         return
 
@@ -218,7 +202,7 @@ def unmonitor_episodes(episode_ids):
         'monitored': False
     }
     response = requests.put(url, headers=headers, json=payload)
-    if response.status_code in [200, 202]:
+    if response.status_code in {200, 202}:
         print(f"Monitoring status changed successfully for episodes: {episodes_to_unmonitor}")
     else:
         print(f"Error changing monitoring status: {response.status_code}, {response.text}")
@@ -228,13 +212,11 @@ def get_movie_info(tmdb_id, movie_name):
     response = requests.get(url)
     if response.status_code == 200:
         movie_info = response.json()
-        # Ensure movie_info is a list and has at least one element
         if isinstance(movie_info, list) and len(movie_info) > 0:
             # Assuming the first movie in the list is the one you want
-            return movie_info[0]  
-        else:
-            print(f"No movies found for radarr for {movie_name}: {response.status_code}, {response.text}")
-            return None
+            return movie_info[0]
+        print(f"No movies found for radarr for {movie_name}: {response.status_code}, {response.text}")
+        return None
     if response.status_code == 404:
         print(f"{movie_name} Not found in Radarr Library")
     else:
@@ -265,7 +247,7 @@ def unmonitor_movies(radarr_id, movie_name):
             'Content-Type': 'application/json'
         }
         response = requests.put(url, json=payload, headers=headers)
-        if response.status_code in [200, 202]:
+        if response.status_code in {200, 202}:
             print(f"Monitoring status updated for {movie_name}.")
         else:
             print(f"Error updating monitoring status for {movie_name}: {response.status_code}, {response.text}")
@@ -282,13 +264,15 @@ def delete_item(item_id, item_info):
 
 def delete_episode_file(series_id, season_number, episode_number, episode_name, series_name):
     episodes = get_episode_info(series_id, season_number)
-    episode_file_id = None
-    for episode in episodes:
-        if episode['episodeNumber'] == episode_number and episode['hasFile']:
-            episode_file_id = episode['episodeFileId']
-            break
-
-    if episode_file_id:
+    if episode_file_id := next(
+        (
+            episode['episodeFileId']
+            for episode in episodes
+            if episode['episodeNumber'] == episode_number
+            and episode['hasFile']
+        ),
+        None,
+    ):
         delete_url = f"{SONARR_URL}/episodeFile/{episode_file_id}?apikey={SONARR_API_KEY}"
         response = requests.delete(delete_url)
         if response.status_code == 200:
@@ -311,32 +295,23 @@ def main():
     watched_series = get_series_info(EMBY_USER_ID)
     if watched_series and IGNORE_FAVOURITES:
         for item in watched_series:
-            if item['UserData']['IsFavorite'] and item['Type'] == 'Series' and HANDLE_TV:
-                BLACKLISTED_PATHS.append(item['Path'])  # Add the path to blacklisted paths (should capture everything)
-                print(f"Favorite Series: {item['Name']} and adding {item['Path']} to blacklisted paths")
-                BLACKLISTED_TV_SHOWS.append(item['Name'])  # Add the Name to blacklisted Movies (Fall back for if path misses something)
-                print(f"Favorite Series: {item['Name']} and adding to blacklisted TV Shows")
-            elif item['UserData']['IsFavorite'] and item['Type'] == 'Movie' and HANDLE_MOVIES:
-                BLACKLISTED_PATHS.append(item['Path'])  # Add the path to blacklisted paths (should capture everything)
-                print(f"Favorite Movie: {item['Name']} and adding {item['Path']} to blacklisted paths")
-                BLACKLISTED_MOVIES.append(item['Name'])  # Add the Name to blacklisted Movies (Fall back for if path misses something)
-                print(f"Favorite Series: {item['Name']} and adding to blacklisted TV Shows")
-                continue
+            if item['UserData']['IsFavorite']:
+                if item['Type'] == 'Series' and HANDLE_TV:
+                    _extracted_from_main_(item, 'Favorite Series: ', BLACKLISTED_TV_SHOWS)
+                elif item['Type'] == 'Movie' and HANDLE_MOVIES:
+                    _extracted_from_main_(item, 'Favorite Movie: ', BLACKLISTED_MOVIES)
     if not IGNORE_FAVOURITES:
         print("Ignore Favourites set to false, WARNING FAVOURITES WILL BE HANDLED BY THE SCRIPT.")
     # Print statement for Handle TV = False
     if not HANDLE_TV:
         print("Handling of TV shows is disabled, Skipping to Movies.")
-    # Logic for hadnling everything and calling the rest of the funcions
-    #Gets list of watched items from emby
-    watched_items = get_watched_items(EMBY_USER_ID)
-    if watched_items:   # only acts on results from emby in watched items
-        for item in watched_items:         #Loops through each item in watched items and applies the following logic
+    if watched_items := get_watched_items(EMBY_USER_ID):
+        for item in watched_items: #Loops through each item in watched items and applies the following logic
             if item['Type'] == 'Episode' and HANDLE_TV: #Checks if item type is an episode and if handle TV = True
                 series_name = item['SeriesName']
                 episode_name = item['Name']
                 season_number = item['ParentIndexNumber']
-                episode_number = item['IndexNumber']  
+                episode_number = item['IndexNumber']
                 item_info = f"{episode_name} from {series_name}"
                 if item['UserData']['IsFavorite']: #Checks if episode is favourited in emby
                     print(f"Skipping favourite item: {item['Name']}")
@@ -347,12 +322,8 @@ def main():
                 if series_name in BLACKLISTED_TV_SHOWS: # Checks if series name is blacklisted
                     print(f"Skipping blacklisted show: {series_name}")
                     continue  # Skip this iteration if the show is blacklisted
-                # Fetch TVDB ID using series name
-                tvdb_id = get_tvdb_id(series_name, item['Path'])
-                if tvdb_id:
-                    # Fetch Sonarr series ID using TVDB ID
-                    series_id = get_series_id_by_tvdb(tvdb_id)
-                    if series_id:
+                if tvdb_id := get_tvdb_id(series_name, item['Path']):
+                    if series_id := get_series_id_by_tvdb(tvdb_id):
                         # Fetch episode info from Sonarr to get episode ID
                         episode_info = get_episode_info(series_id, season_number)
                         #print(episode_info)
@@ -388,39 +359,45 @@ def main():
                     continue
                 if movie_name in BLACKLISTED_MOVIES:
                     print(f"Skipping blacklisted Movie: {movie_name}")
-                    continue  # Skip this iteration if the Movie is blacklisted
-                elif movie_name not in BLACKLISTED_MOVIES:
-                    # Use the movie name to fetch the TMDB ID directly
-                    tmdb_id = get_tmdb_id(movie_name, item['Path'])
-                    if tmdb_id:
-                        movie_info = get_movie_info(tmdb_id, movie_name)
-                        if movie_info:
-                            radarr_id = movie_info['id']  # Assuming the first movie is the correct one
-                            release_dates = [movie_info.get('inCinemas'), movie_info.get('physicalRelease'), movie_info.get('digitalRelease')]
-                            release_dates = [datetime.strptime(date, '%Y-%m-%dT%H:%M:%SZ') for date in release_dates if date is not None]
-                            if release_dates:  # Ensure there's at least one valid date
-                                earliest_release_date = min(release_dates) if release_dates else None
-                                Do_not_delete = datetime.now() - timedelta(days=DAYS)
-                                # Unmonitor movies in Radarr
-                                unmonitor_movies(radarr_id, movie_name) 
-                                print(f"Unmonitored movie: {movie_name} with TMDB ID {tmdb_id}")
-                                if earliest_release_date < Do_not_delete and MOVIE_DELETE:
-                                    #delete the episode from Emby(OPTIONAL), Radarr and file system
-                                    delete_movie_file(radarr_id, movie_name)
-                                    if EMBY_DELETE:
-                                        delete_item(item['Id'], item_info)
-                                    if not EMBY_DELETE:
-                                        print(f"Emby library update handled by Sonarr skipping Emby library delete for {movie_name}")
-                                else:
-                                    print(f"{movie_name} aired within the past {DAYS} days. Not Deleted")
+                elif tmdb_id := get_tmdb_id(movie_name, item['Path']):
+                    if movie_info := get_movie_info(tmdb_id, movie_name):
+                        radarr_id = movie_info['id']  # Assuming the first movie is the correct one
+                        release_dates = [movie_info.get('inCinemas'), movie_info.get('physicalRelease'), movie_info.get('digitalRelease')]
+                        if release_dates := [
+                            datetime.strptime(date, '%Y-%m-%dT%H:%M:%SZ')
+                            for date in release_dates
+                            if date is not None
+                        ]:
+                            earliest_release_date = min(release_dates, default=None)
+                            Do_not_delete = datetime.now() - timedelta(days=DAYS)
+                            # Unmonitor movies in Radarr
+                            unmonitor_movies(radarr_id, movie_name)
+                            print(f"Unmonitored movie: {movie_name} with TMDB ID {tmdb_id}")
+                            if earliest_release_date < Do_not_delete and MOVIE_DELETE:
+                                #delete the episode from Emby(OPTIONAL), Radarr and file system
+                                delete_movie_file(radarr_id, movie_name)
+                                if EMBY_DELETE:
+                                    delete_item(item['Id'], item_info)
+                                if not EMBY_DELETE:
+                                    print(f"Emby library update handled by Sonarr skipping Emby library delete for {movie_name}")
                             else:
-                                print(f"No valid release dates available for comparison: {movie_name}")
-                    else:
-                        print(f"TMDB ID not found for movie '{item_info}'.")
+                                print(f"{movie_name} aired within the past {DAYS} days. Not Deleted")
+                        else:
+                            print(f"No valid release dates available for comparison: {movie_name}")
+                else:
+                    print(f"TMDB ID not found for movie '{item_info}'.")
     else:
         print("No watched items found from Emby.")
     if not HANDLE_MOVIES:
         print("Handling of movies is disabled, Ending script.")
+
+
+# TODO Rename this here and in `main`
+def _extracted_from_main_(item, arg1, arg2):
+    BLACKLISTED_PATHS.append(item['Path'])  # Add the path to blacklisted paths (should capture everything)
+    print(f"{arg1}{item['Name']} and adding {item['Path']} to blacklisted paths")
+    arg2.append(item['Name'])
+    print(f"Favorite Series: {item['Name']} and adding to blacklisted TV Shows")
 
 if __name__ == "__main__":
     main()
